+++ src/models/TimelineEvents.ts
@@ -4,14 +4,14 @@ import {
   GetRegisteredServiceProviderEventsResponse,
   GetDeregisteredServiceProviderEventsResponse,
   GetIncreasedStakeEventResponse
-} from '../services/Audius/serviceProviderClient'
+} from '../services/Audius/service-provider/types'
 import { GetIncreaseDelegateStakeEventsResponse } from 'services/Audius/delegate/types'
 import {
   GetClaimEventsResponse,
   GetDelegatorRemovedEventsResponse,
   GetSlashEventsResponse
 } from 'services/Audius/delegate/types'
-import { GetClaimProcessedResponse } from '../services/Audius/claim'
+import { GetClaimProcessedResponse } from '../services/Audius/claim/types'
 
 /* ServiceProvider Events */
