+++ src/services/Audius/claim/types.ts
@@ -1,0 +1,10 @@
+import { Address } from 'types'
+import BN from 'bn.js'
+
+export type GetClaimProcessedResponse = {
+  blockNumber: number
+  claimer: Address
+  rewards: BN
+  oldTotal: BN
+  newTotal: BN
+}
